// src/api/webform/api.post.webform.type.ts
export type T_FormGetInvitedRequest = {
  webform_id: string;
  nama: string;
  email: string;
  telepon: string;
  apakah_anda_nasabah_bri: string;
  metode_kontak: string;
  waktu_dihubungi: string;
  saya_ingin: string;
  pesan: string;
  pilih_provinsi: string;
  pilih_lokasi: string;
};

export type T_FormResult = {
  sid: string;
};